Genisis/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
comtypes==1.1.7
docopt==0.6.2
idna==2.10
Js2Py==0.70
more-itertools==8.6.0
packaging==20.8
pafy==0.5.5
pip==20.2.3
pipwin==0.5.0
PyAudio==0.2.11
pyjsparser==2.7.1
pyparsing==2.4.7
pypiwin32==223
PyPrind==2.11.2
pySmartDL==1.3.4
python-vlc==3.0.11115
pyttsx3==2.90
pytz==2020.4
pywikipedia==0.0.1
pywin32==300
requests==2.25.1
setuptools==49.2.1
six==1.15.0
soupsieve==2.1
SpeechRecognition==3.8.1
tzlocal==2.1
urllib3==1.26.2
wikipedia==1.4.0
wolframalpha==4.1.1
xmltodict==0.12.0
youtube-dl==2020.12.14
